requirements.txt
influxdb
python-crontab
pyyaml
python-dotenv